Fetch redirect URLs from referrer chain as setup for redirect-related signals
Add a function that will derive a set of unique redirect URLs to provide more information that's helpful in determining if a site is suspicious. This patch provides the setup for signals related to the redirect chain, such as whether the site redirected through a TLD associated with a high rate of spam/abuse.

Part of the work for issues #22, #23, and #24

PiperOrigin-RevId: 263251521