Fix broken caching
Store the URL for the MostRecent alerts list, and use it to verify that the list is appropriate to use on a subsequent query.

Also, change to a persistent background script so the data (e.g. topsites list, most recent alerts) isn't blown away.

Also, change from using chrome.extension.connect to chrome.runtime.sendMessage.