Remove top sites alert and make gray flag the no-alert state
The top sites alert on its own doesn't seem to be particularly informative as to
whether a site should be reported, so remove this alert. The top sites list still
continues to be used to determine whether to alert on IDNs and redirects from
outside program or webmail.

Since the gray flag was originally shown when the top sites alert was the only one,
this intermediate state is no longer needed. This change replaces the green flag
no-alert state with gray and removes the green flag in order to provide better
consistency with coloring on other security UX (e.g. gray lock icon).

There was a small correlation between unvisited sites + not top sites to site
badness, so move the notVisitedBefore alert within the !isTopSite conditional.

PiperOrigin-RevId: 309964380